+++ Changelog.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - ```userPermissions``` method throwing error when no submission or submission id is specified
  - ```Formio``` tests always failing with timeout error instead of actual error
  - ```Formio``` tests ignoring mock results for same URLs
+ 
+### Added
+ - Tests for ```userPermissions``` method
 
 ## 4.8.0-beta.5
 ### Fixed